Extract `ModsTable` Component Into Own File
